Starting an implementation for exporting files, not finished/compiling.

+++ app/src/main/java/net/kwatts/powtools/database/AttributeDao.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Dao
 public interface AttributeDao {
     @Query("SELECT * FROM Attribute where :momentId = moment_id")
-    List<Attribute> getFromMoment(int momentId);
+    List<Attribute> getFromMoment(long momentId);
 
     @Query("SELECT * "
             + "FROM Attribute "
